Started adding blob framework
I thought it would be useful to create a map of swathes of vertical /
horizontal line segments of contiguous color, as this would be the
easiest when fitting legos with a width of 1. But I don't think that
will be useful in the long run, and am leaning toward a map of booleans
instead.