Changed row/column iterators to return empties instead onf Nones
Which is the more elegant design choice, but that shouldn't matter at
all for now. In face if I try to get an index out of bounds I want to
know about it because something's wrong. So I'm reverting to the old
system but unwrapping instead of returning Options, and maybe in the
future I'll un-revert this. I have mixed feelings about this because
this is pretty elegant code that ends up looking like dog shit :/